requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.6.0
google-genai==0.3.0
qdrant-client==1.12.0
python-dateutil==2.9.0
tenacity==9.0.0
